# server/openapi/agent-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asha News Agent API",
    "version": "1.1.0",
    "description": "Read-focused endpoints for AI agents and digest consumers.",
    "x-legacy-deprecations-doc": "/API_DEPRECATIONS.md"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Cluster": {
        "type": "object",
        "properties": {
          "id": { "type": ["string", "number"] },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "topic": { "type": "string" },
          "article_count": { "type": "integer" },
          "source_count": { "type": "integer" },
          "created_at": { "type": ["string", "null"], "format": "date-time" },
          "status": { "type": "string" }
        }
      },
      "InstrumentNewsItem": {
        "type": "object",
        "properties": {
          "id": { "type": ["string", "number"] },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "source": { "type": "string" },
          "url": { "type": ["string", "null"], "format": "uri" },
          "category": { "type": ["string", "null"] },
          "published_at": { "type": ["string", "null"], "format": "date-time" },
          "bias": { "type": ["string", "null"] }
        }
      },
      "InstrumentPrice": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string" },
          "quote_symbol": { "type": "string" },
          "price": { "type": ["number", "null"] },
          "change": { "type": ["number", "null"] },
          "change_pct": { "type": ["number", "null"] },
          "currency": { "type": ["string", "null"] },
          "market_state": { "type": ["string", "null"] },
          "as_of": { "type": ["string", "null"], "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/openapi": {
      "get": {
        "summary": "Get this OpenAPI document",
        "responses": {
          "200": {
            "description": "OpenAPI document"
          }
        }
      }
    },
    "/digest": {
      "get": {
        "summary": "Get digest (personal or public)",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "schema": { "type": "string", "enum": ["personal", "public"] }
          },
          {
            "name": "topic",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "source",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "bias",
            "in": "query",
            "schema": { "type": "string", "enum": ["all", "left", "center", "right"] }
          },
          {
            "name": "timeframe",
            "in": "query",
            "schema": { "type": "string", "enum": ["1h", "6h", "24h", "3d", "7d", "30d"] }
          },
          {
            "name": "contentType",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["json", "markdown"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Digest payload",
            "content": {
              "application/json": {
                "examples": {
                  "digestExample": {
                    "value": {
                      "generated_at": "2026-02-28T10:00:00.000Z",
                      "scope": "public",
                      "filters": { "topic": "markets", "timeframe": "24h", "source": "", "bias": "all", "contentType": "all" },
                      "user": null,
                      "clusters": [
                        {
                          "id": 123,
                          "title": "Oil extends gains amid supply concerns",
                          "summary": "Energy markets reacted to fresh supply disruptions.",
                          "topic": "markets",
                          "article_count": 12,
                          "source_count": 9,
                          "created_at": "2026-02-28T09:30:00.000Z",
                          "status": "active"
                        }
                      ],
                      "digest_text": "Asha News public digest (1 clusters)"
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Auth required for personal digest" }
        }
      }
    },
    "/digest/share-token": {
      "post": {
        "summary": "Enable/disable or rotate public digest token for authenticated user",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" },
                  "rotate": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Share token state updated",
            "content": {
              "application/json": {
                "examples": {
                  "enabled": {
                    "value": {
                      "success": true,
                      "enabled": true,
                      "token": "abc123",
                      "public_path": "/api/v1/public/token/abc123"
                    }
                  },
                  "disabled": {
                    "value": {
                      "success": true,
                      "enabled": false
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Authentication required" },
          "404": { "description": "User profile not found" }
        }
      }
    },
    "/public/{userId}": {
      "get": {
        "summary": "Get public digest for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "Public digest payload" },
          "403": { "description": "Public digest disabled" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/public/token/{token}": {
      "get": {
        "summary": "Get public digest by share token",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "Public digest payload" },
          "404": { "description": "Token not found" }
        }
      }
    },
    "/clusters/search": {
      "get": {
        "summary": "Search story clusters",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Cluster search results",
            "content": {
              "application/json": {
                "examples": {
                  "clusterSearch": {
                    "value": {
                      "query": "btc",
                      "count": 1,
                      "results": [
                        {
                          "id": 88,
                          "title": "Bitcoin jumps above key resistance",
                          "summary": "Crypto markets rallied after macro data release.",
                          "topic": "markets",
                          "article_count": 7,
                          "source_count": 6,
                          "created_at": "2026-02-28T10:15:00.000Z",
                          "status": "active"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing query" }
        }
      }
    },
    "/instruments/{symbol}/news": {
      "get": {
        "summary": "Get instrument-related news",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Instrument news results",
            "content": {
              "application/json": {
                "examples": {
                  "instrumentNews": {
                    "value": {
                      "symbol": "BTCUSD",
                      "aliases": ["btc", "bitcoin", "btcusd"],
                      "count": 2,
                      "results": [
                        {
                          "id": 1001,
                          "title": "Bitcoin extends weekly rally",
                          "summary": "Analysts cite ETF inflows and risk sentiment.",
                          "source": "Reuters",
                          "url": "https://example.com/article",
                          "category": "markets",
                          "published_at": "2026-02-28T11:00:00.000Z",
                          "bias": "center"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/instruments/prices": {
      "get": {
        "summary": "Get market ticker price snapshot for supported symbols",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Comma-separated symbols, e.g. BTCUSD,ETHUSD,XAUUSD"
          }
        ],
        "responses": {
          "200": {
            "description": "Price snapshot results",
            "content": {
              "application/json": {
                "examples": {
                  "priceSnapshot": {
                    "value": {
                      "source": "yahoo",
                      "updated_at": "2026-02-28T11:01:00.000Z",
                      "count": 2,
                      "results": [
                        {
                          "symbol": "BTCUSD",
                          "quote_symbol": "BTC-USD",
                          "price": 64321.5,
                          "change": 512.3,
                          "change_pct": 0.8,
                          "currency": "USD",
                          "market_state": "REGULAR",
                          "as_of": "2026-02-28T11:00:55.000Z"
                        },
                        {
                          "symbol": "ETHUSD",
                          "quote_symbol": "ETH-USD",
                          "price": 3412.1,
                          "change": -22.7,
                          "change_pct": -0.66,
                          "currency": "USD",
                          "market_state": "REGULAR",
                          "as_of": "2026-02-28T11:00:55.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}